stegapp/backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Pillow>=10.0.0
numpy>=1.24.0,<2.0.0
PyPDF2==3.0.1
cryptography>=41.0.0
pytest>=7.4.0
pytest-cov>=4.1.0
ruff>=0.1.0
